requirements.txt
flask
celery
redis
flask-sqlalchemy
Flask-Migrate
requests
python-dotenv